+++ packages/ionic/src/commands/cordova/prepare.ts
@@ -61,7 +61,11 @@ export class PrepareCommand extends CordovaCommand implements CommandPreRun {
       const platformEngines = await conf.getPlatformEngines();
 
       if (platformEngines.length === 0) {
-        this.env.log.warn(`No Cordova platforms listed in ${chalk.bold('config.xml')}. Nothing to prepare.`);
+        this.env.log.warn(
+          `No Cordova platforms listed in ${chalk.bold('config.xml')}. Nothing to prepare.\n` +
+          `You can save your installed platforms to ${chalk.bold('config.xml')} with the ${chalk.green('ionic cordova platform save')} command.`
+        );
+
         return 0;
       }
     }